Added initial Blocks and Bounces modules
I want to be a hard-arse and drop support for ruby 1.9 so the
gem can make use shiny new ruby 2 features. Keyword arguments seem
fitting as a way to enforce the allowed arguments for the api.

All required parameters should be specified in a similar way but
manually throw an ArgumentError if nil.