Added many more API endpoints...
Filters, invalid emails, mail and credentials.

API methods now return a Sendgrid::Web::Response which correctly
parses the response body using Oj.

Faraday was switched out as the test stubs were not working as
required. Now HTTParty is used and the internal connection method
points to a private "API" class that includes HTTParty.

A shared spec example was created to make testing endpoints easier
as well as reducing the amount of duplication when stubbing the
connection.