Big documentation update using yard.
Fixed a regression caused by switching to HTTParty which meant
parameters weren't being included in query string. I have updated
the specs and turned `default_params` into a method that accepts a
hash. The provided hash is then merged nto the query string as is
expected. All previous merges are now using the method.

On a side-note I have began trying to enforce a 70 character width
in-order to help keep the code style consistent.